1.로그인 및 회원 가입 폼에서 브라우저 경고 뜨는 것 수정 2. 프로필 html에 모달 추가 3. 프로필 js에 비밀 번호 변경 ajax 및 회원 탈퇴 시 서명 추가

+++ src/main/resources/static/js/profile/profile.js
@@ -84,6 +84,17 @@ const WithDrawlManipulator = {
         }
     },
     withdraw: function () {
+
+        const withdrawalConfirmPlaceHolder=$("#withdrawConfirm").attr('placeholder');
+        const withdrawalConfirm=$("#withdrawConfirm").val();
+
+        console.log(withdrawalConfirm,withdrawalConfirmPlaceHolder);
+
+        if(withdrawalConfirm!==withdrawalConfirmPlaceHolder){
+            swal('', "자신의 이메일을 제대로 서명해주세요", 'error');
+            return;
+        }
+
         $.ajax({
             type: 'DELETE',
             url: "/profile/withdrawal",
@@ -105,6 +116,10 @@ const UpdatePasswordManipulator = {
             _this.changeVisibilityOfModal();
         });
 
+        $("#updatePasswordModalDecideBtn").on('click', function () {
+            _this.changePassword();
+        });
+
         $("#updatePasswordModalCloseBtn").on('click', function () {
             _this.changeVisibilityOfModal();
         });
@@ -117,6 +132,35 @@ const UpdatePasswordManipulator = {
             $("#updatePasswordModal").attr("style", "display:none;");
             this.isModalOn = false;
         }
+    },
+
+    /*
+    흠... 아래 메서드는 보안 요소 추가가 필요할 듯..?
+     */
+    changePassword: function () {
+        const data = {
+            password: $("#passwordInput").val(),
+            passwordConfirm: $("#passwordConfirmInput").val()
+        }
+
+        if (data.password.length < 8 || data.password.length > 20 || data.passwordConfirm.length < 8 || data.passwordConfirm.length > 20) {
+            swal('', "변경하고자 하는 비밀 번호는 길이가 8이상 20이하여야 합니다.", 'error');
+            return;
+        }
+
+        $.ajax({
+            type: 'PUT',
+            url: '/profile/password',
+            dataType: 'json',
+            contentType: 'application/json; charset=utf-8',
+            data: JSON.stringify(data)
+        }).done(function (e) {
+            if (e.success === true) {
+                swal('', e.response, 'success');
+            } else {
+                swal('', e.error.message, 'error');
+            }
+        });
     }
 }
